Refactoring the execution step to some utils and some model methods
